cli/requirements.txt
grpcio==1.7.3
cliff==2.9.1
